+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/ContainedFluid.java
@@ -1,0 +1,9 @@
+package com.refinedmods.refinedstorage2.platform.common;
+
+import com.refinedmods.refinedstorage2.api.resource.ResourceAmount;
+import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
+
+import net.minecraft.world.item.ItemStack;
+
+public record ContainedFluid(ItemStack remainderContainer, ResourceAmount<FluidResource> fluid) {
+}