+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/resource/ResourceContainerImpl.java
@@ -267,6 +267,7 @@ public class ResourceContainerImpl implements ResourceContainer {
         for (int i = 0; i < size(); ++i) {
             final String key = "s" + i;
             if (!tag.contains(key)) {
+                removeSilently(i);
                 continue;
             }
             final CompoundTag item = tag.getCompound(key);