+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/render/model/AbstractFluidRenderer.java
@@ -7,8 +7,12 @@ import com.mojang.blaze3d.vertex.BufferBuilder;
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.Tesselator;
+import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.blaze3d.vertex.VertexFormat;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 
 public abstract class AbstractFluidRenderer implements FluidRenderer {
@@ -43,5 +47,48 @@ public abstract class AbstractFluidRenderer implements FluidRenderer {
             .uv(sprite.getU0(), sprite.getV0())
             .color(r, g, b, 255).endVertex();
         tesselator.end();
+    }
+
+    protected void render(final PoseStack poseStack,
+                          final MultiBufferSource renderTypeBuffer,
+                          final int light,
+                          final int packedRgb,
+                          final TextureAtlasSprite sprite) {
+        final VertexConsumer buffer = renderTypeBuffer.getBuffer(RenderType.text(sprite.atlasLocation()));
+        final float scale = 0.3F;
+        // y is flipped here
+        final var x0 = -scale / 2;
+        final var y0 = scale / 2;
+        final var x1 = scale / 2;
+        final var y1 = -scale / 2;
+        final var transform = poseStack.last().pose();
+        buffer.vertex(transform, x0, y1, 0)
+            .color(packedRgb)
+            .uv(sprite.getU0(), sprite.getV1())
+            .overlayCoords(OverlayTexture.NO_OVERLAY)
+            .uv2(light)
+            .normal(0, 0, 1)
+            .endVertex();
+        buffer.vertex(transform, x1, y1, 0)
+            .color(packedRgb)
+            .uv(sprite.getU1(), sprite.getV1())
+            .overlayCoords(OverlayTexture.NO_OVERLAY)
+            .uv2(light)
+            .normal(0, 0, 1)
+            .endVertex();
+        buffer.vertex(transform, x1, y0, 0)
+            .color(packedRgb)
+            .uv(sprite.getU1(), sprite.getV0())
+            .overlayCoords(OverlayTexture.NO_OVERLAY)
+            .uv2(light)
+            .normal(0, 0, 1)
+            .endVertex();
+        buffer.vertex(transform, x0, y0, 0)
+            .color(packedRgb)
+            .uv(sprite.getU0(), sprite.getV0())
+            .overlayCoords(OverlayTexture.NO_OVERLAY)
+            .uv2(light)
+            .normal(0, 0, 1)
+            .endVertex();
     }
 }